Dev to main (#24)
* [#6] primary and secondary navbar

* [#10] Palette component

Palette Component with page to this functionalitie and Main Layout for all app's pages

* [#12] created tile component 

Now in project we have TileComponent.jsx

* [#13] Mosaic component

Mosaic Component - third element needed to showing colors in this aplication. This is the smallest one which functionalitie is to just show the color and contains button to preview of containing color.

* [#16] Routing for three colors views

Added routing with react-routing-v6. Changes in NavButton which IS the main component to navigate in application.
Plus some small changes - just details.

* hotfix - detail comment

* [#19] - Created ColorsContext

* [#21] New Readme

Update README.md